Better data type

+++ app/database/migrations/2013_05_07_215025_create_values_table.php
@@ -15,13 +15,13 @@ class CreateValuesTable extends Migration {
             $table->increments('id');
             $table->integer('stock_id')->unsigned();
             $table->date('date');
-            $table->integer('open');
-            $table->integer('close');
-            $table->integer('high');
-            $table->integer('low');
+            $table->decimal('open', 9, 5);
+            $table->decimal('close', 9, 5);
+            $table->decimal('high', 9, 5);
+            $table->decimal('low', 9, 5);
             $table->integer('volume')->unsigned();
-            $table->integer('delta');
-            $table->integer('absdelta');
+            $table->decimal('delta', 9, 5);
+            $table->decimal('absdelta', 9, 5);
             $table->foreign('stock_id')->references('id')->on('stocks');
         });
     }
